requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-login
python-dotenv
weasyprint
brasilapi
pytest
pytest-flask
